requirements.txt
autopep8==1.6.0
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
Flask==2.1.1
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
pycodestyle==2.8.0
python-dotenv==0.20.0
requests==2.27.1
toml==0.10.2
urllib3==1.26.9
Werkzeug==2.1.1
